move `printout` function to lib.py
